% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stprob.R
\name{as.stprob}
\alias{as.stprob}
\alias{as.stprob.array}
\title{Coerce an object to a \code{stprob} object}
\usage{
as.stprob(x, ...)

\method{as.stprob}{array}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Ignored.}
}
\value{
A \code{stprob} object containing the state transition probabilities.
}
\description{
Currently only works with objects obtained from \link[TraMineR:seqtrate]{TraMineR::seqtrate}.
}
\section{Methods (by class)}{
\itemize{
\item \code{as.stprob(array)}: Convert the result of \link[TraMineR:seqtrate]{TraMineR::seqtrate} to a
\code{stprob} object.

}}
