% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradu-package.R
\docType{package}
\name{gradu}
\alias{gradu}
\title{The \code{gradu} package.}
\description{
An user friendly interface to estimating state transition probabilities
using a Markov model. Includes plotting tools to visualize the probabilities
over time between groups.
}
\section{See also}{
\itemize{
\item \link{stprobs} for information on estimating state transition probabilities.
\item \link{stprops} for information on calculating state transition proportions.
\item \link{plot.stprob} for information on plotting the probabilities.
\item \href{https://github.com/joakim219/gradu/issues/}{Submit a bug report or a feature request}
}
}

