% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhmgrid-package.R
\docType{data}
\name{health}
\alias{health}
\title{Simulated health state dataset}
\format{
\subsection{\code{health}}{

A data frame with 500 rows and 4 columns:
\describe{
\item{id}{Person identification number}
\item{age}{Person age in years}
\item{state}{Measured health state: healthy, sick or deceased}
\item{lagstate}{Previous measured health state}
}
}
}
\usage{
health
}
\description{
This panel data contains simulated health state measurements from 50
individuals over 10 years. This data set is used in the examples of the
\link{nhmgrid} package.
}
\keyword{datasets}
