% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct.R
\name{manual_struct}
\alias{manual_struct}
\title{TODO document}
\usage{
manual_struct(state, x, group = NULL, prob, ...)
}
\arguments{
\item{state}{description}

\item{x}{description}

\item{group}{description}

\item{prob}{description}

\item{...}{description}
}
\description{
TODO document
}
