% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhmgrid-package.R
\docType{package}
\name{nhmgrid}
\alias{nhmgrid}
\title{The \code{nhmgrid} package.}
\description{
A user friendly interface to estimating state transition probabilities
using a Markov model. Includes plotting tools to visualize the probabilities
over time between groups.
}
\section{See also}{
\itemize{
\item \link{health} for information on the data set used in examples.
\item \link{stprobs} for information on estimating state transition probabilities.
\item \link{stprops} for information on calculating state transition proportions.
\item \link{plot.stprob} for information on plotting the probabilities.
\item \link{manual_stprob} for information on creating a \code{stprob} object manually.
\item \href{https://github.com/joakim219/nhmgrid/issues/}{Submit a bug report or a feature request}
}
}

