% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.stprob}
\alias{plot.stprob}
\title{Plot State Transition Probabilities}
\usage{
\method{plot}{stprob}(x, default_geoms = TRUE, ...)
}
\arguments{
\item{x}{[\code{nhmgrid::stprob}]\cr
The \code{stprob} object containing the transition probabilities.}

\item{default_geoms}{[\code{logical(1)}]\cr
See 'Details' and 'Examples'.}

\item{...}{Ignored.}
}
\description{
Plot State Transition Probabilities
}
\details{
The function defined in \code{fun_cell} is applied to all cells of the plot matrix
individually. The function should be defined with 1 or 2 arguments and it
should return the modified plot. The first argument is assigned the original
plot object obtained with \link[ggplot2:ggplot]{ggplot2::ggplot}. If the second argument is
defined, it contains the index of the cell in the plot matrix. See 'Examples'.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Estimate state transition probabilities in the health data and plot them
fit <- nnet::multinom(state ~ lagstate + age, nhmgrid::health)
probs <- nhmgrid::stprobs(fit, x = "age")
plot(probs)

# Calculate state transition proportions
props <- nhmgrid::stprops(nhmgrid::health, "id", "state", "age")
plot(props)

# The argument `fun_cell` can be used to customize the individual cells # TODO

# One could add a smoothing lines in all of the plots
plot(props, fun_cell = \(cell) cell + ggplot2::geom_smooth())

# or only in a specific cell by using a 2 argument function
plot(props, fun_cell = \(cell, index) {
  if (index == 5) {
    return(cell + ggplot2::geom_smooth())
  }
  return(cell)
})
\dontshow{\}) # examplesIf}
}
