% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stprob.R
\name{stprobs}
\alias{stprobs}
\title{State Transition Probabilities}
\usage{
stprobs(model, x = NULL, group = NULL, lag_state = NULL, interval = 0.95)
}
\arguments{
\item{model}{The model object.}

\item{x}{[\code{character(1)}]\cr
The predictor name in \code{model} which represents time.}

\item{group}{[\code{character(1)}]\cr
The predictor name in \code{model} which to group the observations by. If \code{NULL},
the observations will not be grouped.}

\item{lag_state}{[\code{character(1)}]\cr
The predictor name in \code{model} which represents lagged states. This should
only be supplied if automatic detection fails.}

\item{interval}{[\code{numeric(1)}]\cr
The confidence/credibility interval.}
}
\value{
A \code{stprob} object containing the estimated state transition probabilities.
}
\description{
Estimate state transition probabilities for given Markov model.
}
\details{
The transition probabilities are estimated using
\link[marginaleffects:predictions]{marginaleffects::avg_predictions}. In case of a \link[dynamite:dynamite]{dynamite::dynamite} model, a custom
implementation of predicting the probabilities is used.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Fit a multinomial logistic regression model
fit <- nnet::multinom(state ~ lagstate + age + sex, nhmgrid::health)

# Estimate and plot the transition probabilities
probs <- nhmgrid::stprobs(fit, x = "age")
plot(probs)

# Estimate and plot the transition probabilities separately for men and women
probs <- nhmgrid::stprobs(fit, x = "age", group = "sex")
plot(probs) +
  ggplot2::labs(title = "Men and women separately")
\dontshow{\}) # examplesIf}
}
