% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stprob.R
\name{stprops}
\alias{stprops}
\title{State Transition Proportions}
\usage{
stprops(data, id, state, x, group = NULL)
}
\arguments{
\item{data}{The data object.}

\item{id}{[\code{character(1)}]\cr
The column name in \code{data} which represents individual identifiers.}

\item{state}{[\code{character(1)}]\cr
The column name in \code{data} which represents states.}

\item{x}{[\code{character(1)}]\cr
The column name in \code{data} which represents time.}

\item{group}{[\code{character(1)}]\cr
The column name in \code{data} which to group the observations by. If \code{NULL}, the
observations will not be grouped.}
}
\value{
A \code{stprob} object containing the calculated state transition proportions.
}
\description{
Calculate state transition proportions for given data.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Data structure
str(nhmgrid::health)

# Calculate and plot empirical transition probabilities
props <- nhmgrid::stprops(nhmgrid::health, "id", "state", "age")
plot(props)

# Group by sex
props <- nhmgrid::stprops(nhmgrid::health, "id", "state", "age", "sex")
plot(props)
\dontshow{\}) # examplesIf}
}
