% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct.R
\name{struct.dynamitefit}
\alias{struct.dynamitefit}
\title{TODO document}
\usage{
\method{struct}{dynamitefit}(
  fit,
  state_name,
  x_values = NULL,
  group = NULL,
  fixed_predictors = NULL,
  interval = NULL,
  ...
)
}
\arguments{
\item{fit}{description}

\item{state_name}{description}

\item{x_values}{description}

\item{group}{description}

\item{fixed_predictors}{description}

\item{interval}{description}

\item{...}{description}
}
\description{
TODO document
}
