% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct.R
\name{struct.manual}
\alias{struct.manual}
\title{TODO document}
\usage{
\method{struct}{manual}(state, x, group = NULL, prob)
}
\arguments{
\item{state}{description}

\item{x}{description}

\item{group}{description}

\item{prob}{description}
}
\description{
TODO document
}
