% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct.R
\name{struct.multinom}
\alias{struct.multinom}
\title{TODO document}
\usage{
\method{struct}{multinom}(
  fit,
  state_name,
  x,
  group = NULL,
  fixed_predictors = NULL,
  interval = NULL,
  ...
)
}
\arguments{
\item{fit}{description}

\item{state_name}{description}

\item{x}{description}

\item{group}{description}

\item{fixed_predictors}{description}

\item{interval}{description}

\item{...}{description}
}
\description{
TODO document
}
