% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct.R
\name{struct}
\alias{struct}
\title{A generic method to estimate transition probabilities.}
\usage{
struct(
  fit,
  x = NULL,
  state = NULL,
  group = NULL,
  fixed_predictors = NULL,
  interval = NULL,
  ...
)
}
\arguments{
\item{fit}{The model object.}

\item{x}{See 'See also'.}

\item{state}{See 'See also'.}

\item{group}{[\verb{character(1) or list(name, values)}]\cr
Either the name of the column to be used as the grouping variable or
a named list of the column name and the desired grouping values.
If \code{NULL}, the observations will not be grouped.}

\item{fixed_predictors}{[\code{list()}]\cr
A named list of parameters to control the other predictors used in the model.}

\item{interval}{[\code{numeric(1)}]\cr
The level of confidence/credible interval. Ignore at the moment.}

\item{...}{Ignored.}
}
\value{
A \code{struct} object which is a list containing the following components:
\itemize{
\item \code{state}\cr A list of the state variable name and values.
\item \code{x}\cr A list of the x-axis variable name and values.
\item \code{group}\cr A list of the grouping variable name and values.
\item \code{prob}\cr A \code{data.table} object of the transition probabilities between the states.
}
}
\description{
Estimates the transition probabilities within groups over time using
a Markov model.
}
\details{
Please view the 'Details' of a specific extended \code{struct} method in 'See also'.
All of the extension methods are listed in the 'See Also' section.
}
\seealso{
Other struct: 
\code{\link{struct.dynamitefit}()},
\code{\link{struct.multinom}()}
}
\concept{struct}
