% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stprob.R
\name{manual_stprob}
\alias{manual_stprob}
\title{Manual creation of a \code{stprob} object}
\usage{
manual_stprob(prob)
}
\arguments{
\item{prob}{[\verb{data.table or data.frame}]\cr
A data table/frame containing the state transition probabilities.
See 'Details'.}
}
\description{
Manual creation of a \code{stprob} object
}
