% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.stprob}
\alias{plot.stprob}
\title{Plot State Transition Probabilities}
\usage{
\method{plot}{stprob}(x, default_geoms = TRUE, ...)
}
\arguments{
\item{x}{[\code{nhmgrid::stprob}]\cr
The \code{stprob} object containing the transition probabilities.}

\item{default_geoms}{[\code{logical(1)}]\cr
Inclusion of default geometric layers. See 'Details' and 'Examples'.}

\item{...}{Ignored.}
}
\value{
A \link{ggplot2} figure.
}
\description{
Plot state transition probabilities as a matrix of figures.
}
\details{
The resulting plot is created with \link{ggplot2}. The matrix-like structure is
constructed with \link[ggplot2:facet_grid]{ggplot2::facet_grid}. By default, certain geometric layers
are added in the plot to visualize the data. In order to replace the default
geometric components one has to define \code{default_geoms} as \code{FALSE} and add the
new layers manually.

When plotting the result of \link{stprobs}, the default geometric components
are \link[ggplot2:geom_path]{ggplot2::geom_line} and \link[ggplot2:geom_ribbon]{ggplot2::geom_ribbon} for the estimates and their
corresponding confidence/credibility intervals. When plotting the result of
\link{stprops}, the default geometric components for the resulting
lollipop plot are \link[ggplot2:geom_point]{ggplot2::geom_point} and \link[ggplot2:geom_segment]{ggplot2::geom_segment}.

One can add any \link{ggplot2} objects to the result of this function. See the
'Examples' section.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Estimate state transition probabilities in the health data and plot them
fit <- nnet::multinom(state ~ lagstate + age, nhmgrid::health)
probs <- nhmgrid::stprobs(fit, x = "age")
plot(probs)

# Calculate state transition proportions and plot them with a custom title
props <- nhmgrid::stprops(nhmgrid::health, "id", "state", "age")
plot(props) +
  ggplot2::labs(title = "Custom title")

# The parameter `default_geoms` can be used to hide/replace the default
# geometric components

# Defining it as FALSE will result in a matrix of empty plots
plot(probs, default_geoms = FALSE)

# Combining this with specific components will replace the default components
plot(probs, default_geoms = FALSE) +
  ggplot2::geom_area()

# One could plot the probabilities and proportions together
plot(probs) +
  ggplot2::geom_point(data = props) +
  ggplot2::geom_smooth(data = props, se = FALSE)
\dontshow{\}) # examplesIf}
}
