% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.stprob}
\alias{plot.stprob}
\title{TODO document}
\usage{
\method{plot}{stprob}(x, fun_cell = identity, ...)
}
\arguments{
\item{x}{description}

\item{fun_cell}{description}

\item{...}{description}
}
\description{
TODO document
}
